+++ src/ansys/health/heart/post/dpf_utils.py
@@ -32,10 +32,13 @@ import pyvista as pv
 
 from ansys.dpf import core as dpf
 from ansys.health.heart import LOG as LOGGER
-from ansys.health.heart.exceptions import SupportedDPFServerNotFoundError
+from ansys.health.heart.exceptions import (
+    MissingEnvironmentVariableError,
+    SupportedDPFServerNotFoundError,
+)
 from ansys.health.heart.models import HeartModel
 
-_SUPPORTED_DPF_SERVERS = ["2025.2", "2024.1", "2024.1rc1", "2024.2rc0"]
+_SUPPORTED_DPF_SERVERS = ["2025.2", "2025.2rc0", "2024.1", "2024.1rc1", "2024.2rc0"]
 """List of supported DPF servers."""
 #! NOTE:
 #! 2024.1rc0: not supported due to missing ::tf operator
@@ -48,7 +51,7 @@ def _check_accept_dpf():
             """DPF requires you to accept the license agreement.
             Set the environment variable "ANSYS_DPF_ACCEPT_LA" to "Y"."""
         )
-        exit()
+        raise MissingEnvironmentVariableError("ANSYS_DPF_ACCEPT_LA not set to 'Y'.")
     return
 
 
@@ -72,7 +75,7 @@ def _get_dpf_server():
         LOGGER.error(mess)
         raise SupportedDPFServerNotFoundError(mess)
 
-    return server
+    return server, version
 
 
 class D3plotReader:
@@ -90,7 +93,7 @@ class D3plotReader:
         _check_accept_dpf()
 
         # TODO: retrieve version from docker
-        self._server = _get_dpf_server()
+        self._server, self._dpf_version = _get_dpf_server()
 
         self.ds = dpf.DataSources()
         self.ds.set_result_file_path(path, "d3plot")
@@ -238,7 +241,13 @@ class D3plotReader:
 
         res = []
         for i in hv_index:
-            res.append(hist_vars[i].data)
+            if self._dpf_version.startswith("2025.2"):
+                # Skip duplicate values.
+                data = hist_vars[i].data[0::3]
+            else:
+                data = hist_vars[i].data
+
+            res.append(data)
 
         return np.array(res)
 
@@ -280,7 +289,10 @@ class ICVoutReader:
             self._get_available_ids()
         except IndexError as error:
             LOGGER.error(f"{fn} does not contain icvout. {error}")
-            exit()
+            raise IndexError(
+                f"File {fn} does not contain control volume data. "
+                "Make sure the binout file contains ICVOUT results."
+            ) from error
 
     def _get_available_ids(self) -> np.ndarray:
         """Get available CV IDs and CVI IDs."""