+++ src/ansys/health/heart/simulator.py
@@ -144,7 +144,10 @@ class BaseSimulator:
             for name in ["alpha", "beta", "beta_septum"]:
                 if name not in rotation_angles.keys():
                     LOGGER.error(f"Must provide key {name} for D-RBM method.")
-                    exit()
+                    raise KeyError(
+                        f"Must provide key {name} for D-RBM method. "
+                        "Please check the settings or provide the rotation angles."
+                    )
 
             self._compute_fibers_lsdyna(rotation_angles)
 
@@ -156,12 +159,18 @@ class BaseSimulator:
             for a, b in zip(["alpha", "beta"], ["_left", "_right", "_ot"]):
                 if a + b not in rotation_angles.keys():
                     LOGGER.error(f"Must provide key {name} for D-RBM method.")
-                    exit()
+                    raise KeyError(
+                        f"Must provide key {name} for D-RBM method. "
+                        "Please check the settings or provide the rotation angles."
+                    )
             self._compute_fibers_drbm(rotation_angles)
 
         else:
             LOGGER.error(f"Method {method} is not recognized.")
-            exit()
+            raise ValueError(
+                f"Method {method} is not recognized. "
+                "Please use 'LSDYNA' or 'D-RBM' as the method for computing fibers."
+            )
 
         return
 
@@ -653,7 +662,7 @@ class MechanicsSimulator(BaseSimulator):
         folder_name="zeropressure",
         overwrite: bool = True,
         extra_k_files: list[str] | None = None,
-    ):
+    ) -> tuple[dict, np.ndarray, np.ndarray]:
         """Compute the stress-free configuration of the model.
 
         Parameters
@@ -664,6 +673,13 @@ class MechanicsSimulator(BaseSimulator):
             Whether to run simulation and overwrite files.
         extra_k_files : list[str], default: None
             User-defined k files.
+
+        Returns
+        -------
+        tuple[dict, np.ndarray, np.ndarray]
+            Dictionary with convergence information,
+            stress free configuration, and
+            (re)computed end-of-diastole configuration.
         """
         directory = os.path.join(self.root_directory, folder_name)
 
@@ -693,7 +709,7 @@ class MechanicsSimulator(BaseSimulator):
                 if isinstance(value, SurfaceMesh):
                     part.__setattr__(key, self.model.mesh.get_surface(value.id))
 
-        return
+        return report, stress_free_coord, guess_ed_coord
 
     def _write_main_simulation_files(
         self,